+++ R/pca_sparse_bayes.R
@@ -257,3 +257,9 @@ tidy.step_pca_sparse_bayes <- function(x, ...) {
   res$id <- x$id
   res
 }
+
+#' @rdname required_pkgs.embed
+#' @export
+required_pkgs.step_pca_sparse_bayes <- function(x, ...) {
+  c("embed", "VBsparsePCA")
+}