+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * All steps now have `required_pkgs()` methods.
 
+* Add `metric` argument to `step_umap()`. (#154)
+
 # embed 1.0.0
 
 * `step_collapse_cart()` can pool a predictor's factor levels using a tree-based method.